fix start times for "deaths since 1/million"
our attempt to find the first date at which there was
1 death/million happend AFTER stat-specific filtering -
which meant that the first day data was available also
became the first day that deaths/million appeared to meet
our 1/million mark.  That is, for graphs with incomplete
data, the line was always drawn starting at day 0.

However, this is not correct - if deaths/million hit 1
on March 1st, but then we didn't start getting
hospitalization data until March 15th, there should be
two weeks of empty space before the hospitalization graph
starts.